Split db.hpp into leveldb_db.hpp and lmdb_db.hpp
